Add new pragma to force rustup to be rust for the build.
A few builds have recently been added which need a newer rust and rely
on the fact that rustup will auto-download a newer rust if needed.

This is not perfect - it really only make sense when using chroots,
which is the default, but it is working (as long as you don't have
any links for rustup in your /usr/local/sbin already).