Add a new pragma `download_urldecode` to adjust DOWNLOAD.
Some slackbuilds might have a url-encoded DOWNLOAD configured but we
might need to use curl to download. In this case, the slackbuild itself
though might expect the url-decoded name to be used for the resulting
file. This combination is impossible to get working with `curl` or `wget
--content-disposition`.

This pragma can run url-decode across the url before using it. You might
also need to combine it with download_basename to force the filename to
be the name expected by the slackbuild when the build runs.

Closes #1